pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "hyperliquid-trading-bot"
version = "0.1.0"
description = "Extensible trading bot for Hyperliquid DEX"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hyperliquid-python-sdk>=0.20.0",
    "eth-account>=0.10.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0",
    "psutil>=7.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "websockets>=15.0.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=43.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "aiosqlite>=0.20.0",
    "numba>=0.60",
    "uvloop>=0.21.0; sys_platform != 'win32'",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.0",
]
dev = [
    "black",
    "isort",
    "mypy",
]
